requirements.txt
waitress==0.8.10
psycopg2==2.6.1
PyYAML==3.11
requests==2.9.1
